+++ lib/src/config/dictionaries.rs
@@ -101,9 +101,7 @@ mod deserialization {
     use {
         super::{DictionariesConfig, Dictionary, DictionaryName},
         crate::{
-            config::limits::{
-                DICTIONARY_ITEM_KEY_MAX_LEN, DICTIONARY_ITEM_VALUE_MAX_LEN, DICTIONARY_MAX_LEN,
-            },
+            config::limits::{DICTIONARY_ITEM_KEY_MAX_LEN, DICTIONARY_ITEM_VALUE_MAX_LEN},
             error::{DictionaryConfigError, FastlyConfigError},
         },
         std::{collections::HashMap, path::PathBuf, str::FromStr},
@@ -225,11 +223,6 @@ mod deserialization {
         dict: &HashMap<String, String>,
     ) -> Result<(), DictionaryConfigError> {
         info!("checking if dictionary adheres to Fastly's API",);
-        if dict.len() > DICTIONARY_MAX_LEN {
-            return Err(DictionaryConfigError::DictionaryCountTooLong {
-                size: DICTIONARY_MAX_LEN.try_into().unwrap(),
-            });
-        }
 
         for (key, value) in dict.iter() {
             if key.chars().count() > DICTIONARY_ITEM_KEY_MAX_LEN {
@@ -252,16 +245,8 @@ mod deserialization {
     impl FromStr for DictionaryName {
         type Err = DictionaryConfigError;
         fn from_str(name: &str) -> Result<Self, Self::Err> {
-            // Name must start with alphabetical and contain only alphanumeric, underscore, and whitespace
-            if name.starts_with(char::is_alphabetic)
-                && name
-                    .chars()
-                    .all(|c| char::is_alphanumeric(c) || c == '_' || char::is_whitespace(c))
-            {
-                Ok(Self(name.to_owned()))
-            } else {
-                Err(DictionaryConfigError::InvalidName(name.to_owned()))
-            }
+            // We do not do any validation on the name as Config Stores and Dictionaries have different validation rules
+            Ok(Self(name.to_owned()))
         }
     }
 }