+++ lib/src/config/object_store.rs
@@ -1,19 +1,19 @@
 use {
     crate::{
         error::{FastlyConfigError, ObjectStoreConfigError},
-        object_store::{ObjectKey, ObjectStore, ObjectStoreKey},
+        object_store::{ObjectKey, ObjectStoreKey, ObjectStores},
     },
     std::fs,
     toml::value::Table,
 };
 
 #[derive(Clone, Debug, Default)]
-pub struct ObjectStoreConfig(pub(crate) ObjectStore);
+pub struct ObjectStoreConfig(pub(crate) ObjectStores);
 
 impl TryFrom<Table> for ObjectStoreConfig {
     type Error = FastlyConfigError;
     fn try_from(toml: Table) -> Result<Self, Self::Error> {
-        let obj_store = ObjectStore::new();
+        let obj_store = ObjectStores::new();
         for (store, items) in toml.iter() {
             let items = items.as_array().ok_or_else(|| {
                 FastlyConfigError::InvalidObjectStoreDefinition {
@@ -39,6 +39,7 @@ impl TryFrom<Table> for ObjectStoreConfig {
                         err: ObjectStoreConfigError::NotATable,
                     }
                 })?;
+
                 let key = item
                     .get("key")
                     .ok_or_else(|| FastlyConfigError::InvalidObjectStoreDefinition {
@@ -50,24 +51,33 @@ impl TryFrom<Table> for ObjectStoreConfig {
                         name: store.to_string(),
                         err: ObjectStoreConfigError::KeyNotAString,
                     })?;
-                let bytes = match (item.get("path"), item.get("data")) {
+
+                // Previously the "file" key was named "path".  We want
+                // to continue supporting the old name.
+                let file = match (item.get("file"), item.get("path")) {
+                    (None, None) => None,
+                    (Some(file), _) => Some(file),
+                    (None, Some(path)) => Some(path),
+                };
+
+                let bytes = match (file, item.get("data")) {
                     (None, None) => {
                         return Err(FastlyConfigError::InvalidObjectStoreDefinition {
                             name: store.to_string(),
-                            err: ObjectStoreConfigError::NoPathOrData(key.to_string()),
+                            err: ObjectStoreConfigError::NoFileOrData(key.to_string()),
                         })
                     }
                     (Some(_), Some(_)) => {
                         return Err(FastlyConfigError::InvalidObjectStoreDefinition {
                             name: store.to_string(),
-                            err: ObjectStoreConfigError::PathAndData(key.to_string()),
+                            err: ObjectStoreConfigError::FileAndData(key.to_string()),
                         })
                     }
                     (Some(path), None) => {
                         let path = path.as_str().ok_or_else(|| {
                             FastlyConfigError::InvalidObjectStoreDefinition {
                                 name: store.to_string(),
-                                err: ObjectStoreConfigError::PathNotAString(key.to_string()),
+                                err: ObjectStoreConfigError::FileNotAString(key.to_string()),
                             }
                         })?;
                         fs::read(path).map_err(|e| {
@@ -86,6 +96,7 @@ impl TryFrom<Table> for ObjectStoreConfig {
                         .as_bytes()
                         .to_vec(),
                 };
+
                 obj_store
                     .insert(
                         ObjectStoreKey::new(store),