+++ lib/src/error.rs
@@ -92,6 +92,9 @@ pub enum Error {
     #[error(transparent)]
     ObjectStoreError(#[from] crate::object_store::ObjectStoreError),
 
+    #[error(transparent)]
+    SecretStoreError(#[from] crate::wiggle_abi::SecretStoreError),
+
     #[error{"Expected UTF-8"}]
     Utf8Expected(#[from] std::str::Utf8Error),
 
@@ -118,6 +121,18 @@ pub enum Error {
 
     #[error("Invalid Object Store `key` value used: {0}.")]
     ObjectStoreKeyValidationError(#[from] crate::object_store::KeyValidationError),
+
+    #[error("Unfinished streaming body")]
+    UnfinishedStreamingBody,
+
+    #[error("Shared memory not supported yet")]
+    SharedMemory,
+
+    #[error("Value absent from structure")]
+    ValueAbsent,
+
+    #[error("String conversion error")]
+    ToStr(#[from] http::header::ToStrError),
 }
 
 impl Error {
@@ -130,7 +145,7 @@ impl Error {
     pub fn to_fastly_status(&self) -> FastlyStatus {
         match self {
             Error::BufferLengthError { .. } => FastlyStatus::Buflen,
-            Error::InvalidArgument => FastlyStatus::Inval,
+            Error::InvalidArgument | Error::ValueAbsent => FastlyStatus::Inval,
             Error::Unsupported { .. } => FastlyStatus::Unsupported,
             Error::HandleError { .. } => FastlyStatus::Badf,
             Error::InvalidStatusCode { .. } => FastlyStatus::Inval,
@@ -145,6 +160,7 @@ impl Error {
             Error::DictionaryError(e) => e.to_fastly_status(),
             Error::GeolocationError(e) => e.to_fastly_status(),
             Error::ObjectStoreError(e) => e.into(),
+            Error::SecretStoreError(e) => e.into(),
             // All other hostcall errors map to a generic `ERROR` value.
             Error::AbiVersionMismatch
             | Error::BackendUrl(_)
@@ -169,7 +185,10 @@ impl Error {
             | Error::BackendNameRegistryError(_)
             | Error::HttpError(_)
             | Error::UnknownObjectStore(_)
-            | Error::ObjectStoreKeyValidationError(_) => FastlyStatus::Error,
+            | Error::ObjectStoreKeyValidationError(_)
+            | Error::UnfinishedStreamingBody
+            | Error::SharedMemory
+            | Error::ToStr(_) => FastlyStatus::Error,
         }
     }
 
@@ -219,6 +238,10 @@ pub enum HandleError {
     #[error("Invalid pending request handle: {0}")]
     InvalidPendingRequestHandle(crate::wiggle_abi::types::PendingRequestHandle),
 
+    /// A lookup handle was not valid.
+    #[error("Invalid pending KV lookup handle: {0}")]
+    InvalidPendingKvLookupHandle(crate::wiggle_abi::types::PendingKvLookupHandle),
+
     /// A dictionary handle was not valid.
     #[error("Invalid dictionary handle: {0}")]
     InvalidDictionaryHandle(crate::wiggle_abi::types::DictionaryHandle),
@@ -226,6 +249,18 @@ pub enum HandleError {
     /// An object-store handle was not valid.
     #[error("Invalid object-store handle: {0}")]
     InvalidObjectStoreHandle(crate::wiggle_abi::types::ObjectStoreHandle),
+
+    /// A secret store handle was not valid.
+    #[error("Invalid secret store handle: {0}")]
+    InvalidSecretStoreHandle(crate::wiggle_abi::types::SecretStoreHandle),
+
+    /// A secret handle was not valid.
+    #[error("Invalid secret handle: {0}")]
+    InvalidSecretHandle(crate::wiggle_abi::types::SecretHandle),
+
+    /// An async item handle was not valid.
+    #[error("Invalid async item handle: {0}")]
+    InvalidAsyncItemHandle(crate::wiggle_abi::types::AsyncItemHandle),
 }
 
 /// Errors that can occur in a worker thread running a guest module.
@@ -243,7 +278,7 @@ pub(crate) enum ExecutionError {
     ///
     /// [call]: https://docs.rs/wasmtime/latest/wasmtime/struct.Func.html#method.call
     #[error("WebAssembly execution trapped: {0}")]
-    WasmTrap(wasmtime::Trap),
+    WasmTrap(anyhow::Error),
 
     /// Errors thrown when trying to instantiate a guest context.
     #[error("Error creating context: {0}")]
@@ -295,6 +330,13 @@ pub enum FastlyConfigError {
         name: String,
         #[source]
         err: ObjectStoreConfigError,
+    },
+
+    #[error("invalid configuration for '{name}': {err}")]
+    InvalidSecretStoreDefinition {
+        name: String,
+        #[source]
+        err: SecretStoreConfigError,
     },
 
     /// An error that occurred while deserializing the file.
@@ -368,9 +410,6 @@ pub enum DictionaryConfigError {
 
     #[error("definition was not provided as a TOML table")]
     InvalidEntryType,
-
-    #[error("invalid string: {0}")]
-    InvalidName(String),
 
     #[error("'name' field was not a string")]
     InvalidNameEntry,
@@ -487,14 +526,14 @@ pub enum ObjectStoreConfigError {
     /// An I/O error that occured while reading the file.
     #[error(transparent)]
     IoError(std::io::Error),
-    #[error("The `path` and `data` keys for the object `{0}` are set. Only one can be used.")]
-    PathAndData(String),
-    #[error("The `path` or `data` key for the object `{0}` is not set. One must be used.")]
-    NoPathOrData(String),
+    #[error("The `file` and `data` keys for the object `{0}` are set. Only one can be used.")]
+    FileAndData(String),
+    #[error("The `file` or `data` key for the object `{0}` is not set. One must be used.")]
+    NoFileOrData(String),
     #[error("The `data` value for the object `{0}` is not a string.")]
     DataNotAString(String),
-    #[error("The `path` value for the object `{0}` is not a string.")]
-    PathNotAString(String),
+    #[error("The `file` value for the object `{0}` is not a string.")]
+    FileNotAString(String),
     #[error("The `key` key for an object is not set. It must be used.")]
     NoKey,
     #[error("The `key` value for an object is not a string.")]
@@ -507,6 +546,39 @@ pub enum ObjectStoreConfigError {
     ObjectStoreError(#[from] crate::object_store::ObjectStoreError),
     #[error("Invalid `key` value used: {0}.")]
     KeyValidationError(#[from] crate::object_store::KeyValidationError),
+}
+
+/// Errors that may occur while validating secret store configurations.
+#[derive(Debug, thiserror::Error)]
+pub enum SecretStoreConfigError {
+    /// An I/O error that occured while reading the file.
+    #[error(transparent)]
+    IoError(std::io::Error),
+
+    #[error("The `file` and `data` keys for the object `{0}` are set. Only one can be used.")]
+    FileAndData(String),
+    #[error("The `file` or `data` key for the object `{0}` is not set. One must be used.")]
+    NoFileOrData(String),
+    #[error("The `data` value for the object `{0}` is not a string.")]
+    DataNotAString(String),
+    #[error("The `file` value for the object `{0}` is not a string.")]
+    FileNotAString(String),
+
+    #[error("The `key` key for an object is not set. It must be used.")]
+    NoKey,
+    #[error("The `key` value for an object is not a string.")]
+    KeyNotAString,
+
+    #[error("There is no array of objects for the given store.")]
+    NotAnArray,
+    #[error("There is an object in the given store that is not a table of keys.")]
+    NotATable,
+
+    #[error("Invalid secret store name: {0}")]
+    InvalidSecretStoreName(String),
+
+    #[error("Invalid secret name: {0}")]
+    InvalidSecretName(String),
 }
 
 /// Errors related to the downstream request.