+++ lib/src/wiggle_abi/geo_impl.rs
@@ -71,7 +71,10 @@ impl FastlyGeo for Session {
         let result_len =
             u32::try_from(result.len()).expect("smaller than value_max_len means it must fit");
 
-        let mut buf_ptr = buf.as_array(result_len).as_slice_mut()?;
+        let mut buf_ptr = buf
+            .as_array(result_len)
+            .as_slice_mut()?
+            .ok_or(Error::SharedMemory)?;
         buf_ptr.copy_from_slice(result.as_bytes());
         nwritten_out.write(result_len)?;
         Ok(())